+++ src/hooks/useNavigateTo.ts
@@ -84,13 +84,6 @@ export default function useNavigateTo() {
           pathname: APP_PATHS.ACCELERATOR_OVERVIEW,
           search: 'tab=projects',
         }),
-
-      goToParticipantProjectSpecies: (deliverableId: number, projectId: number, speciesId: number) =>
-        navigate({
-          pathname: APP_PATHS.ACCELERATOR_SPECIES.replace(':speciesId', `${speciesId}`)
-            .replace(':projectId', `${projectId}`)
-            .replace(':deliverableId', `${deliverableId}`),
-        }),
     }),
     [navigate]
   );
