+++ src/providers/ParticipantProject/ParticipantProjectSpeciesContext.tsx
@@ -1,10 +1,13 @@
 import { createContext, useContext } from 'react';
 
 import { ParticipantProjectSpecies } from 'src/services/ParticipantProjectSpeciesService';
+import { Deliverable } from 'src/types/Deliverables';
 
 export type ParticipantProjectSpeciesData = {
   participantProjectSpecies: ParticipantProjectSpecies[];
   currentParticipantProjectSpecies?: ParticipantProjectSpecies;
+  currentDeliverable?: Deliverable;
+  setCurrentDeliverable: (deliverable: Deliverable) => void;
   setCurrentParticipantProjectSpecies: (projectSpeciesId: string | number) => void;
   reload: () => void;
 };
@@ -14,6 +17,7 @@ export const ParticipantProjectSpeciesContext = createContext<ParticipantProject
   // tslint:disable-next-line:no-empty
   participantProjectSpecies: [],
   setCurrentParticipantProjectSpecies: () => {},
+  setCurrentDeliverable: () => {},
   reload: () => {},
 });
