+++ src/scenes/DeliverablesRouter/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 
-import DeliverableProvider from 'src/providers/Deliverable/DeliverableProvider';
 import ParticipantProvider from 'src/providers/Participant/ParticipantProvider';
 import ProjectProvider from 'src/providers/Project/ProjectProvider';
 
@@ -17,11 +16,9 @@ const DeliverablesRouter = (): JSX.Element => {
         element={
           <ProjectProvider>
             <ParticipantProvider>
-              <DeliverableProvider>
-                <ParticipantProjectProvider>
-                  <DeliverableViewWrapper />
-                </ParticipantProjectProvider>
-              </DeliverableProvider>
+              <ParticipantProjectProvider>
+                <DeliverableViewWrapper />
+              </ParticipantProjectProvider>
             </ParticipantProvider>
           </ProjectProvider>
         }
